# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NovaDev Cloud IDE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 2 */
  username: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  pro: 'pro',
  team: 'team',
  enterprise: 'enterprise',
} as const;

export interface User {
  id: number;
  email: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  plan?: UserPlan;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  archived: 'archived',
  deleted: 'deleted',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  language: string;
  userId: number;
  isPublic?: boolean;
  status?: ProjectStatus;
  /** @nullable */
  lastOpenedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectList {
  projects: Project[];
  total: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  language: string;
  isPublic?: boolean;
  /** @nullable */
  templateId?: number | null;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  name?: string;
  description?: string;
  isPublic?: boolean;
  status?: ProjectUpdateStatus;
}

export type ProjectStatsLanguageBreakdownItem = {
  language: string;
  count: number;
};

export interface ProjectStats {
  totalProjects: number;
  recentlyActive: number;
  languageBreakdown: ProjectStatsLanguageBreakdownItem[];
  totalDeployments: number;
  storageUsedMb?: number;
}

export type FileNodeType = typeof FileNodeType[keyof typeof FileNodeType];


export const FileNodeType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileNode {
  id: number;
  name: string;
  path: string;
  type: FileNodeType;
  projectId: number;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  size?: number | null;
  children?: FileNode[];
}

export type FileContentType = typeof FileContentType[keyof typeof FileContentType];


export const FileContentType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileContent {
  id: number;
  name: string;
  path: string;
  type: FileContentType;
  content: string;
  projectId: number;
  /** @nullable */
  language?: string | null;
}

export type FileInputType = typeof FileInputType[keyof typeof FileInputType];


export const FileInputType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileInput {
  name: string;
  path: string;
  type: FileInputType;
  content?: string;
  /** @nullable */
  parentId?: number | null;
}

export interface FileUpdate {
  content?: string;
  name?: string;
}

export type AiMessageRole = typeof AiMessageRole[keyof typeof AiMessageRole];


export const AiMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface AiMessage {
  id: number;
  role: AiMessageRole;
  content: string;
  projectId: number;
  createdAt: string;
}

export interface AiMessageInput {
  message: string;
  context?: string;
  model?: string;
}

export type DeploymentStatus = typeof DeploymentStatus[keyof typeof DeploymentStatus];


export const DeploymentStatus = {
  pending: 'pending',
  building: 'building',
  running: 'running',
  failed: 'failed',
  stopped: 'stopped',
} as const;

export type DeploymentEnvironment = typeof DeploymentEnvironment[keyof typeof DeploymentEnvironment];


export const DeploymentEnvironment = {
  production: 'production',
  staging: 'staging',
  preview: 'preview',
} as const;

export interface Deployment {
  id: number;
  projectId: number;
  status: DeploymentStatus;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  buildLog?: string | null;
  environment?: DeploymentEnvironment;
  createdAt: string;
  updatedAt?: string;
}

export type DeploymentInputEnvironment = typeof DeploymentInputEnvironment[keyof typeof DeploymentInputEnvironment];


export const DeploymentInputEnvironment = {
  production: 'production',
  staging: 'staging',
  preview: 'preview',
} as const;

export type DeploymentInputEnvVars = {[key: string]: string};

export interface DeploymentInput {
  environment: DeploymentInputEnvironment;
  envVars?: DeploymentInputEnvVars;
}

export interface Template {
  id: number;
  name: string;
  language: string;
  description: string;
  category?: string;
  tags?: string[];
  starterFiles?: number;
}

export type ListProjectsParams = {
search?: string;
language?: string;
limit?: number;
offset?: number;
};