# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NovaDev Cloud IDE API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{id}/files:
    get:
      operationId: getFileTree
      tags: [files]
      summary: Get project file tree
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileNode"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file or directory
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "201":
          description: File created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileNode"
  /projects/{id}/files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file content
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
    patch:
      operationId: updateFile
      tags: [files]
      summary: Update file content
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects/{id}/deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List project deployments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Deploy a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Deployment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /projects/{id}/deployments/{deployId}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: deployId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterInput:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        username:
          type: string
          minLength: 2
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, username, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, pro, team, enterprise]
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    FileNode:
      type: object
      required: [id, name, path, type, projectId]
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        projectId:
          type: integer
        parentId:
          type: ["integer", "null"]
        language:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        children:
          type: array
          items:
            $ref: "#/components/schemas/FileNode"
    FileContent:
      type: object
      required: [id, name, path, content, type, projectId]
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
        projectId:
          type: integer
        language:
          type: ["string", "null"]
    FileInput:
      type: object
      required: [name, path, type]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
        parentId:
          type: ["integer", "null"]
    FileUpdate:
      type: object
      properties:
        content:
          type: string
        name:
          type: string
...
    Deployment:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        status:
          type: string
          enum: [pending, building, running, failed, stopped]
        url:
          type: ["string", "null"]
        buildLog:
          type: ["string", "null"]
        environment:
          type: string
          enum: [production, staging, preview]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeploymentInput:
      type: object
      required: [environment]
      properties:
        environment:
          type: string
          enum: [production, staging, preview]
        envVars:
          type: object
          additionalProperties:
            type: string
...