# lib/api-zod/src/generated/types/deploymentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NovaDev Cloud IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { DeploymentInputEnvironment } from './deploymentInputEnvironment';
import type { DeploymentInputEnvVars } from './deploymentInputEnvVars';

export interface DeploymentInput {
  environment: DeploymentInputEnvironment;
  envVars?: DeploymentInputEnvVars;
}