# lib/api-zod/src/generated/types/fileContent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NovaDev Cloud IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { FileContentType } from './fileContentType';

export interface FileContent {
  id: number;
  name: string;
  path: string;
  type: FileContentType;
  content: string;
  projectId: number;
  /** @nullable */
  language?: string | null;
}